Split robot load into it's own plan (#491)
* Move fill in energy to a util function

* Split robot load into its own plan

* Remove references to visit_directory

* Rename plans to be more explicit about what they do

* Remove accidentally added file
