+++ tests/unit_tests/hyperion/external_interaction/test_ispyb_utils.py
@@ -4,7 +4,6 @@ import pytest
 
 from mx_bluesky.hyperion.external_interaction.callbacks.common.ispyb_mapping import (
     get_proposal_and_session_from_visit_string,
-    get_visit_string_from_path,
 )
 from mx_bluesky.hyperion.external_interaction.ispyb.ispyb_utils import (
     get_current_time_string,
@@ -18,24 +17,6 @@ def test_get_current_time_string():
 
     assert isinstance(current_time, str)
     assert re.match(TIME_FORMAT_REGEX, current_time) is not None
-
-
-@pytest.mark.parametrize(
-    "visit_path, expected_match",
-    [
-        ("/dls/i03/data/2022/cm6477-45/", "cm6477-45"),
-        ("/dls/i03/data/2022/cm6477-45", "cm6477-45"),
-        ("/dls/i03/data/2022/mx54663-1/", "mx54663-1"),
-        ("/dls/i03/data/2022/mx54663-1", "mx54663-1"),
-        ("/dls/i03/data/2022/mx53-1/", None),
-        ("/dls/i03/data/2022/mx53-1", None),
-        ("/dls/i03/data/2022/mx5563-1565/", None),
-        ("/dls/i03/data/2022/mx5563-1565", None),
-    ],
-)
-def test_find_visit_in_visit_path(visit_path: str, expected_match: str):
-    test_visit_path = get_visit_string_from_path(visit_path)
-    assert test_visit_path == expected_match
 
 
 @pytest.mark.parametrize(
